+++ packages/nextjs-template/utils/useFuse.ts
@@ -29,11 +29,7 @@ function useFuse(
   const [loading, setLoading] = useState<boolean>(false);
   const [fuse, setFuse] = useState<FuseNote>();
   const dendronConfig = useDendronConfig();
-  const fuzzThreshold = (
-    dendronConfig
-      ? ConfigUtils.getLookup(dendronConfig)
-      : genDefaultLookupConfig()
-  ).note.fuzzThreshold;
+  const fuzzThreshold = dendronConfig ? ConfigUtils.getLookup(dendronConfig).note.fuzzThreshold : 0.6;
 
   useEffect(() => {
     if (_.isUndefined(fuse)) {